buildworld: Specify cpp for rpcgen to use rather than fallback
Without the -Y switch, rpcgen will fall back to "cpp" which is guided
by OBJFORMAT environment settings to point at the host cpp from gcc50.
In the later stages of buildworld, we want to use the cpp we just built,
not the host cpp which could have issues.  By stage 4, no native host
tools should be in use.