Restore ability for master to build release 4.2
It turns out that setting values on /etc/defaults/make.conf is a bad
idea.  In this case, the two SENDMAIL values that were set prevent
a master branch from building previous releases.

The fix is to comment out those two definitions and move the default
values to the makefile that uses them, using the "?=" operator.