i386 removal, part 11/x: Remove wrong machine/ setup in the boot Makefiles.
Several of the link targets were pointing to ${.CURDIR}/../../i386/...
which doesn't exist. This is from times when our platform-specific
headers were differently laid out and it was never noticed that after
some point, it was no longer needed.

In fact, sys/boot builds properly without all this, because:

* -m32 ensures that the i386 platform is targeted by the compiler

* Various (needed) __i386__ checks in sys/cpu/x86_64/include ensure that
  even on x86_64, the basic types have the right sizes.

Confirmed by binary compare of /boot without and with the patch.