Sync ACPICA code with the Intel licensed 20150619 distribution.
Due to some issue in Intel's release procedure, the dual licensed
distribution - which we use - had not just license differences but
also code differences when comparing to the Intel licensed one.
Adjust to the code of the Intel licensed distribution.

Said issue will be reported to Intel.

This fixes a build issue in emulators/qemu-devel where iasl(8) was
wrongly crashing due to those code differences.

Reported-by: marino