gcc5: Fix quickworld.
When copying the Makefile from gcc47, the writing of a comment line to
configargs.h about it being a generated file was removed, but the output
redirection was not adjusted to create the file from scratch.

This led to quickworld just appending to the existing header in /usr/obj
and resultant build errors.