Import OpenBSD's libm (trunk, 4 July 2015) to a new vendor branch
With the exception of x86_64 (as amd64), all arch-specific files have
been filtered out.  A new directory called "include" has been added
to capture global and arch-specific headers for installation.  Other
than that directory and those exclusions, this is an unaltered copy
of OpenBSD's libm as it existed on 4th of July, 2015.