WPA Supplicant: Add warning about its use
The wpa_supplicant software in base is three releases behind and likely
full of security holes.  For example, the current release in DPorts has
already had 5 security vulnerabilities patched, but the base version is
not being actively maintained.

The base wpa_supplicant should only be used long enough to be able to
install the latest version from dports.  In order to let sysadmins know
that running it is not the best idea, a 10-second warning will come up
when the base wpa_supplicant is used.

reviewed by: robgar