OpenLIBM: Try to fix nextafterl
My first attempt at fixing this function to deal with the always-true
<unsigned variable> > 0 condition was a mistake.  The problem was that
esy and esy were typed as u_int32_t when they should have been "int"
according to math_private.h.

This still doesn't fix the rounding error in Fortran so it's possible
this function is still broken (upstream too).  The FreeBSD version is
a bit different.