OpenLIBM: Replace nextafterl function with FreeBSD's version
The nextafterl function is not coming up with the correct results.  On
GCC's fortran round_4 test, a bit was being added to the lx component
where on FreeBSD it was being subtracted.

Replacing the source file with the equivalent of FreeBSD's version fixed
the rounding regression we saw after moving to OpenBSD's libm.  Upstream
will be informed, so potentially this is only temporary.