OpenLIBM: Replace complex ACOSH functions
The regression tests revealed a bug in the OpenBSD complex COSH
function implementation.  The FreeBSD implementation has been brought in,
at least temporarily, to address the regression.

It is likely that the ASINH functions are deficient with the edge cases
as well, but they've been left alone for now.  The OpenBSD complex
functions implementation probably needs to be reviewed and reworked based
on my review of the OpenSolaris libm and FreeBSD msun implementations