Create short name locales
In other (non-BSD?) platforms, it's possible to use a locale such as
"en_GB".  Up until now, this did not work on DragonFly.  It required
a codeset as well, so if "en_GB.UTF-8" wasn't specified, setlocale
would fail.

Every locale except two has a version that uses the UTF-8 codeset,
so a symlink was created to these to create these territory locales.
The other two are
  hi_IN => hi_IN.ISCII-DEV
  la_LN => la_LN.ISO8859-1

This should improve cross-platform compatibility, and it will lead
to a change in gcc50 libstdc++ locale handling.

While here, set the "share" directory to build in parallel rather
than serially.