locales: Add 66 "@euro" symlinks, and UTF8 alias
While not documented, it appears that modifiers work.  The only widely
used modifier is "@euro" which is an alias for ISO 8859-15 coded which
contains the Euro currently along with 7 other new symbols as compared
to ISO 8859-1.  For all 33 localization sets that use la_LN.ISO8859-15
for LC_CTYPE, I'm creating an @euro and .ISO-8859-15@euro symlink to
them.  These are still found on Linux and the .ISO-8859* extension is
the chosen linux format (this is unfortanately not standardized thus
BSD and Linux differ).  Thus any use of @euro will probably also refer
to the Linux format and that's why I chose it.

For the same reason, .UTF8 versions (Linux) are being symlinked to
DragonFly's .UTF-8 versions.  Symlinks are cheap.