Move some locales before upcoming overhauls
Several locales are obsolete and have misleading names.  This moves
many of them at once.  An entry to the UPDATING files notifies what
the new locale options are for users of affected locales.

This is necessary because the upcoming locale updates are going to be
generated and we can avoid carrying ancient aliases over.  Also note
that Serbia's change may not be complete until the entire locales are
updated.