cldr2def tool: Implement c_format generation
The time definitions were missing c_fmt.  The patch I applied to make
the tool run caused it (it was always a bug, but I didn't fix it
correctly).  The c_formats are not defined, so I'm taking the full
date-time format and stripping any %z or %Z timezone symbols that are
contained within.  It's not perfect, but not too bad either considering.