cldr2def tool: Fix alternative month generation
For some locales defined in tools/tools/locale/etc/charmaps.xml,
there are alternative month names defined in CDATA.  When they are
not present, the POSIX months are just repeated, otherwise this CDATA
set is returned.

The problem is that the spaces and approximately half of the double
quotes were retained as part of the names.  This modification will
strip off the white space and quotes to fix alternative month
definitions.