localedef(1): Better bootstrap fix
We don't need to build localedef in ctools as well as bootstrap tools.
Moreover, when I removed from bootstrap tools, it built but segfaulted
on use.  The COLL_WEIGHTS_MAX file needs to be in the header because
it affects the structure definition.  When it's moved there, the build
on older platforms will succeed.