locales: Change defaults for territory-only locales
From research, it appears that the default for territory-only locales
(e.g. en_US, de_DE) is to use ISO-8859-X (where X is not 15).  The
@euro modifier is the alias for ISO-8859-15.

Since there is no apparent standard, I am going to switch the defaults
from aliasing to UTF-8 to ISO-8859 where available.  For the non-latin
character sets, I left these at UTF-8 rather than try to decide on a
different default.

Incidentally, this enables a number of gcc libstdc++ tests to pass as
well.