i386 removal, part 26/x: Remove i386 specific lib/libc/quad.
None of the platforms which lead to it being compiled into FreeBSD's
libc look too interesting, so remove it.

Also remove copies of a number of those functions residing in libkern
and stop compiling one of these files into libstand. Libstand also
has a quad.h header. Shrink it together to what is actually needed
(not much), while here.