Remove kk_KZ.PT154 and all latin locales
Latin, strangely enough, is not covered by CLDR POSIX files.  Thus,
it is impossible to completely load the locales.  Both LC_CTYPE and
LC_COLLATE need updated files.  Thus, all version of the Latin
locales needs to go.

The kk_KZ locale just came back with the UTF-8 codeset, but there are
character maps for PT154, so it's impossible to continue to have this
locale.