locales: Fix pt_BR.ISO8859-15, remove UTF-8
The pt_BR.ISO8859-15 locale was not properly defined because it's
directory was not pre-created (fixed now).  The UTF-8 CTYPE is no longer
used (every locale explicitly is accounted for) so let's remove it.

The Brazilian locale won't always be fixed under quickworld, it will
require a full build in some cases.