cldr2def: Update Makefile to generate new POSIX source files
We've been using CLDR v2.0.1 because it was the last version that
provided generated POSIX source files.  (The next release, v21, only
provided the java tool to generate them).  The last release is version
27.0.1.  This alternation to the makefile allows the generation of the
desired 72 base locales with a single command (after the tool is
downloaded and installed in the same CLDR release directory)

Note that kk_KZ locale is changing to kk_Cyrl_KZ.

While here, make the necessary updates to charmaps.xml to generate
to locale source files.