groff/mdoc: Disable ru-KOI8-R localization
The template for the ru-KOI8-R localization for groff -mdoc option was
not created in ASCII as it should have been.  Neither was the French
version, which I was able to fix.  Since I can't fix the data that uses
what I believe is an obsolete codeset, I'm disabling it from the build,
effectively removing the locale.

From what I can tell, neither Russian nor French is used on DragonFly
anyway.  It appears to be something inherited from FreeBSD.  I am also
unsure what is special about these two particular mdoc options as they
are the only locales supported.

With this commit, it is possible to buildworld with locale set (e.g.
to en_US.UTF-8, which correctly failed after regex upgrade)