locales: Switch several shortnames away from ISO-8859-1
For all western European countries that have their short names aliased
to ISO-8859-1, this is against POLA.  Once the decision to not use UTF-8
is made, the short names should be linked to ISO-8859-15, the one with
the Euro symbol and other common characters that ISO-8859-1 lacks.

While here, change en_PH to UTF-8 since the Peso sign can't be represented
and do the same for Costa Rica and it's Colon currency.