locales: Add "common.UTF-8" locale
See previous commit message.  This locale is en_US.UTF-8 with a new
LC_CTYPE that covers all common charactersets (not the full UTF-8, but
a very good chunk of it).  This LC_CTYPE source file is manually
created, unfortunately, so it can slowly rot if not updated.

Other locales can use the versatile support of UTF-8 by setting
LC_CTYPE to "common.UTF-8" in the environment.