sbin/hammer: Print total bytes of reblocked nodes
just like it shows both "data elements" and "data bytes"
in the next two lines.

"btree bytes" is kind of obvious since btree nodes are
all 4KB, and there isn't a field for "btree bytes" in
struct hammer_ioc_reblock as well, however this isn't
obvious for users so it should print it.