sys/vfs/hammer: Fix wrong node type in switch/case
A node type is either internal or leaf (or deleted which
is originally internal or leaf), but not record.

A node element's btype could be internal, leaf or record,
but hammer_print_btree_node() is passing a node type to
this function, so cases must be either internal or leaf.

hammer_print_btree_node() and hammer_print_btree_elm()
are not used so nothing changes anyway.