sys/vfs/hammer: Properly access node element base
elm should dereference leaf.base instead of internal.base
since the node that contains node elm here is a leaf node.
These two refer to the same address, but the leaf.base is
the right one in terms of code.

Also print elm->leaf.base.btype which is always expected
to be 'R'.