bump __DragonFly_version for environ, __progname change
The "environ" and "__progname" symbols no longer appear in libc.
Certain linker flags can result in a link failure on these symbols.  This
has been seen in llvm (likely requiring a patch for all versions to treat
DragonFly like FreeBD).  Since this is a behavior change, I think it is
worth a version bump in case makefiles have to check the version to
adjust.