Activate symbol versioning on libc (bump DF version)
The symbol maps for libc have been verified to work as expected, thus
symbol versioning has been activated for libc.  The major version of
libc remains at 8 indefinitely.

The previously built packages and programs should continue to function
normally.  Future packages built on latest master (400304+) will not
run on DragonFly with version 400303 or earlier due to missing symbol
errors by rtld.  The plan is to archive current packages under a new
directory of mirror-master to continue (static) support for older versions
of master until those machines can be upgraded.

The DF version is changed to 400304 with this commit.