installworld: When installing libraries, install libc first
The libraries have to be built in a specific order.  The way the makefile
is written, the libraries are installed in the same order as they are
built.  However, this dynamic made it impossible to install an new libc
that signficantly impacts other libraries in-place.  For example,
libncurses is significantly impacted by libc versioning, and it is
installed before libc which causes immediate failure in programs like
"ln" and "install" because rtld can't find new symbols.

The solution is update the lib makefile to put libc at the beginning
of the installation queue.  This simple change ensures libc changes
will not break world installation.